# Models/ContentModels.cs
namespace personal_website_blazor.Models;

public class PostModel
{
    public string Title { get; set; } = string.Empty;
    public string Description { get; set; } = string.Empty;
    public DateTime? PublishDate { get; set; }
    public DateTime? UpdatedAt { get; set; }
    public string[] Tags { get; set; } = Array.Empty<string>();
    public string[] Technologies { get; set; } = Array.Empty<string>();
    public string[] Topics { get; set; } = Array.Empty<string>();
    public string[] RelatedProjects { get; set; } = Array.Empty<string>();
    public string[] RelatedPosts { get; set; } = Array.Empty<string>();
    public string? CanonicalUrl { get; set; }
    public string ContentType { get; set; } = string.Empty;
    public string Status { get; set; } = "published";
    public string Content { get; set; } = string.Empty;
    public string SearchableText { get; set; } = string.Empty;
    public string Slug { get; set; } = string.Empty;
    public string Section { get; set; } = string.Empty;
    public string Language { get; set; } = "en";
    public string? Image { get; set; }
    public string? Author { get; set; }
    public TocItem[] TocItems { get; set; } = Array.Empty<TocItem>();
}

public class TocItem
{
    public string Id { get; set; } = string.Empty;
    public string Text { get; set; } = string.Empty;
    public int Level { get; set; }
}

public class ContentMetadata
{
    public string Title { get; set; } = string.Empty;
    public string Href { get; set; } = string.Empty;
    public string PublishedAt { get; set; } = string.Empty;
    public string? UpdatedAt { get; set; }
    public string Summary { get; set; } = string.Empty;
    public string SearchableText { get; set; } = string.Empty;
    public string[] Tags { get; set; } = Array.Empty<string>();
    public string[] Technologies { get; set; } = Array.Empty<string>();
    public string[] Topics { get; set; } = Array.Empty<string>();
    public string[] RelatedProjects { get; set; } = Array.Empty<string>();
    public string[] RelatedPosts { get; set; } = Array.Empty<string>();
    public string? CanonicalUrl { get; set; }
    public string ContentType { get; set; } = string.Empty;
    public string Status { get; set; } = "published";
    public string Language { get; set; } = "en";
}

public class SearchResult
{
    public string Title { get; set; } = string.Empty;
    public string Summary { get; set; } = string.Empty;
    public string Href { get; set; } = string.Empty;
    public string TypeLabel { get; set; } = string.Empty;
    public string? PublishedAt { get; set; }
    public string? UpdatedAt { get; set; }
    public string MatchSnippet { get; set; } = string.Empty;
    public string[] Tags { get; set; } = Array.Empty<string>();
}

public class SocialMediaLink
{
    public string[] Type { get; set; } = Array.Empty<string>();
    public string Link { get; set; } = string.Empty;
    public string Label { get; set; } = string.Empty;
    public bool External { get; set; }
    public string Category { get; set; } = "Other";
    public string? IconColor { get; set; }
}