# wwwroot/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Samet Can Cıncık Personal Website API",
    "version": "2.0.0",
    "description": "Public structured access to Samet Can Cıncık's profile, published content, and GitHub repository data."
  },
  "servers": [{ "url": "/", "description": "Same-origin server" }],
  "paths": {
    "/api/content/all": {
      "get": {
        "summary": "List all published content metadata",
        "responses": { "200": { "description": "Content metadata", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ContentMetadata" } } } } } }
      }
    },
    "/api/content/search": {
      "get": {
        "summary": "Search published content",
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string", "maxLength": 200 } },
          { "name": "section", "in": "query", "required": false, "schema": { "type": "string", "enum": ["posts", "gists", "projects"] } }
        ],
        "responses": { "200": { "description": "Search results", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/SearchResult" } } } } } }
      }
    },
    "/api/content/{section}": {
      "get": {
        "summary": "List content in a section",
        "parameters": [{ "$ref": "#/components/parameters/Section" }],
        "responses": { "200": { "description": "Content records", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Post" } } } } }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/api/content/{section}/{slug}": {
      "get": {
        "summary": "Get one content record",
        "parameters": [{ "$ref": "#/components/parameters/Section" }, { "$ref": "#/components/parameters/Slug" }],
        "responses": { "200": { "description": "Content record", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Post" } } } }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/api/content/cv": {
      "get": {
        "summary": "Get the CV as rendered HTML",
        "responses": { "200": { "$ref": "#/components/responses/HtmlDocument" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/api/readme": {
      "get": {
        "summary": "Get the public GitHub profile README as rendered HTML",
        "responses": { "200": { "$ref": "#/components/responses/HtmlDocument" }, "503": { "$ref": "#/components/responses/Unavailable" } }
      }
    },
    "/api/repos": {
      "get": {
        "summary": "List public GitHub repositories",
        "responses": { "200": { "description": "Repository snapshot", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/GitHubRepository" } } } } }, "503": { "$ref": "#/components/responses/Unavailable" } }
      }
    },
    "/api/profile": {
      "get": {
        "summary": "Get the canonical structured author profile",
        "responses": { "200": { "description": "Author profile", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProfileDocument" } } } } }
      }
    },
    "/api/about": {
      "get": {
        "summary": "Get the canonical author profile alias",
        "responses": { "200": { "description": "Author profile", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProfileDocument" } } } } }
      }
    },
    "/api/timeline": {
      "get": {
        "summary": "Get career and education timeline",
        "responses": { "200": { "description": "Timeline entries", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/TimelineEntry" } } } } } }
      }
    },
    "/api/skills": {
      "get": {
        "summary": "Get skills, interests, and languages",
        "responses": { "200": { "description": "Skills document", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SkillsDocument" } } } } }
      }
    },
    "/api/profile/github": {
      "get": {
        "summary": "Get GitHub profile and repository snapshot",
        "responses": { "200": { "description": "GitHub profile document", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GitHubProfileDocument" } } } }, "503": { "$ref": "#/components/responses/Unavailable" } }
      }
    },
    "/health": {
      "get": {
        "summary": "Get application health",
        "responses": { "200": { "description": "Healthy", "content": { "application/json": { "schema": { "type": "object", "properties": { "status": { "type": "string", "example": "ok" } } } } } } }
      }
    }
  },
  "components": {
    "parameters": {
      "Section": { "name": "section", "in": "path", "required": true, "schema": { "type": "string", "enum": ["posts", "gists", "projects"] } },
      "Slug": { "name": "slug", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^[a-z0-9_-]+$" } }
    },
    "responses": {
      "NotFound": { "description": "Resource not found" },
      "Unavailable": { "description": "Upstream data is temporarily unavailable" },
      "HtmlDocument": { "description": "Rendered HTML document", "content": { "application/json": { "schema": { "type": "object", "required": ["html"], "properties": { "html": { "type": "string" } } } } } }
    },
    "schemas": {
      "ContentMetadata": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "href": { "type": "string" },
          "publishedAt": { "type": "string" },
          "updatedAt": { "type": ["string", "null"] },
          "summary": { "type": "string" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "technologies": { "type": "array", "items": { "type": "string" } },
          "topics": { "type": "array", "items": { "type": "string" } },
          "relatedProjects": { "type": "array", "items": { "type": "string" } },
          "relatedPosts": { "type": "array", "items": { "type": "string" } },
          "canonicalUrl": { "type": ["string", "null"] },
          "contentType": { "type": "string" },
          "status": { "type": "string" },
          "language": { "type": "string" }
        }
      },
      "Post": {
        "allOf": [
          { "$ref": "#/components/schemas/ContentMetadata" },
          { "type": "object", "properties": { "content": { "type": "string" }, "slug": { "type": "string" }, "section": { "type": "string" }, "tocItems": { "type": "array", "items": { "$ref": "#/components/schemas/TocItem" } } } }
        ]
      },
      "TocItem": { "type": "object", "properties": { "id": { "type": "string" }, "text": { "type": "string" }, "level": { "type": "integer" } } },
      "SearchResult": { "type": "object", "properties": { "title": { "type": "string" }, "summary": { "type": "string" }, "href": { "type": "string" }, "typeLabel": { "type": "string" }, "publishedAt": { "type": ["string", "null"] }, "updatedAt": { "type": ["string", "null"] }, "matchSnippet": { "type": "string" }, "tags": { "type": "array", "items": { "type": "string" } } } },
      "ProfileLink": { "type": "object", "properties": { "label": { "type": "string" }, "url": { "type": "string" }, "type": { "type": "string" } } },
      "ExperienceEntry": { "type": "object", "properties": { "organization": { "type": "string" }, "role": { "type": "string" }, "location": { "type": "string" }, "startDate": { "type": "string" }, "endDate": { "type": "string" }, "highlights": { "type": "array", "items": { "type": "string" } } } },
      "EducationEntry": { "type": "object", "properties": { "institution": { "type": "string" }, "program": { "type": "string" }, "level": { "type": "string" }, "startDate": { "type": "string" }, "endDate": { "type": "string" }, "highlights": { "type": "array", "items": { "type": "string" } } } },
      "CertificationEntry": { "type": "object", "properties": { "name": { "type": "string" }, "issuer": { "type": "string" }, "url": { "type": ["string", "null"] } } },
      "ProfileDocument": { "type": "object", "properties": { "name": { "type": "string" }, "alternateName": { "type": "string" }, "jobTitle": { "type": "string" }, "location": { "type": "string" }, "summary": { "type": "string" }, "biography": { "type": "string" }, "skills": { "type": "array", "items": { "type": "string" } }, "areasOfInterest": { "type": "array", "items": { "type": "string" } }, "languages": { "type": "array", "items": { "type": "string" } }, "publicNotes": { "type": "array", "items": { "type": "string" } }, "links": { "type": "array", "items": { "$ref": "#/components/schemas/ProfileLink" } }, "experience": { "type": "array", "items": { "$ref": "#/components/schemas/ExperienceEntry" } }, "education": { "type": "array", "items": { "$ref": "#/components/schemas/EducationEntry" } }, "certifications": { "type": "array", "items": { "$ref": "#/components/schemas/CertificationEntry" } }, "lastUpdated": { "type": "string" } } },
      "TimelineEntry": { "type": "object", "properties": { "type": { "type": "string" }, "organization": { "type": "string" }, "role": { "type": "string" }, "level": { "type": "string" }, "location": { "type": "string" }, "startDate": { "type": "string" }, "endDate": { "type": "string" }, "highlights": { "type": "array", "items": { "type": "string" } } } },
      "SkillsDocument": { "type": "object", "properties": { "skills": { "type": "array", "items": { "type": "string" } }, "areasOfInterest": { "type": "array", "items": { "type": "string" } }, "languages": { "type": "array", "items": { "type": "string" } }, "publicNotes": { "type": "array", "items": { "type": "string" } }, "source": { "type": "string" }, "lastUpdated": { "type": "string" } } },
      "GitHubProfile": { "type": "object", "properties": { "login": { "type": "string" }, "name": { "type": ["string", "null"] }, "bio": { "type": ["string", "null"] }, "company": { "type": ["string", "null"] }, "location": { "type": ["string", "null"] }, "blog": { "type": ["string", "null"] }, "avatarUrl": { "type": ["string", "null"] }, "htmlUrl": { "type": "string" }, "publicRepos": { "type": "integer" }, "followers": { "type": "integer" }, "following": { "type": "integer" } } },
      "GitHubRepository": { "type": "object", "properties": { "name": { "type": "string" }, "description": { "type": ["string", "null"] }, "language": { "type": ["string", "null"] }, "fork": { "type": "boolean" }, "archived": { "type": "boolean" }, "stargazersCount": { "type": "integer" }, "forksCount": { "type": "integer" }, "openIssuesCount": { "type": "integer" }, "watchersCount": { "type": "integer" }, "size": { "type": "integer", "format": "int64" }, "homepage": { "type": ["string", "null"] }, "defaultBranch": { "type": ["string", "null"] }, "license": { "type": ["string", "null"] }, "cloneUrl": { "type": ["string", "null"] }, "sshUrl": { "type": ["string", "null"] }, "topics": { "type": "array", "items": { "type": "string" } }, "htmlUrl": { "type": "string" }, "createdAt": { "type": ["string", "null"], "format": "date-time" }, "updatedAt": { "type": ["string", "null"], "format": "date-time" }, "pushedAt": { "type": ["string", "null"], "format": "date-time" } } },
      "GitHubProfileDocument": { "type": "object", "properties": { "profile": { "$ref": "#/components/schemas/GitHubProfile" }, "repositories": { "type": "array", "items": { "$ref": "#/components/schemas/GitHubRepository" } }, "source": { "type": "string" }, "retrievedAt": { "type": "string", "format": "date-time" } } }
    }
  }
}